requirements.txt
google-genai
googlemaps
requests
python-dotenv
reachy-mini
mcp
numpy
sounddevice

fastmcp
